refactor(web): split league home into components/league/* (feedback #11)
Pure moves — route file (1069 -> 138 lines) keeps the query, states, and
canAct composition; header/members/invites/settings/danger-zone each get a
focused module, leagueQueryKey shared from query-key.ts.
